Tones working on Linux and hardware but can hear a slight stuttering on hardware

+++ src/resources/engine_tone_sound_resource.h
@@ -19,7 +19,7 @@
 // ENGINE_AUDIO_SAMPLE_RATE_PERIOD   [second/samples]
 //
 // Therefore: ENGINE_TONE_BUFFER_LEN  =  ENGINE_TONE_MINIMUM_PERIOD / ENGINE_AUDIO_SAMPLE_RATE_PERIOD  =  [second/1.0] / [second/samples]  =  [second/1.0] * [samples/second]  = [samples]
-#define ENGINE_TONE_BUFFER_LEN (ENGINE_TONE_MINIMUM_PERIOD / ENGINE_AUDIO_SAMPLE_RATE_PERIOD)
+#define ENGINE_TONE_BUFFER_LEN ((uint32_t)(ENGINE_TONE_MINIMUM_PERIOD / ENGINE_AUDIO_SAMPLE_RATE_PERIOD))
 
 
 typedef struct{
@@ -27,7 +27,9 @@ typedef struct{
     audio_channel_class_obj_t *channel;
 
     float frequency;
+    // float omega;
     bool busy;
+    // float time;
 
     mp_obj_t data;  // When the tone frequncey is changed, the new waveform data is stored in here
 }tone_sound_resource_class_obj_t;
